pyproject.toml
[project]
name = "api-first"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.6"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "pyyaml>=6.0",
    "email-validator>=2.0.0",
]
